tcp: Extracted ReadRequest from Connection to separate header
